change price to double

+++ src/main/java/com/example/TradingApp/model/Item.java
@@ -1,4 +1,4 @@
 package com.example.TradingApp.model;
 
-public record Item(String title,boolean inMarket,String status,int price,String currency,int discount,String offerId,String type,String category) {
+public record Item(String title, boolean inMarket, String status, double price, String currency, int discount, String offerId, String type, String category) {
 }
